Moved GenerateAST, GetStructFields, and ParseTextAreas to parse.go
